+++ rgen/rgen-jni-impl/benches/chunk.rs
@@ -13,11 +13,11 @@ use test::Bencher;
 fn bench_chunk(b: &mut Bencher) {
   let world = Arc::new(CachedWorld::new());
   let context = Arc::new(rgen_world::Context {
-    seed:   1 as u64,
+    seed:   1_u64,
     blocks: Blocks::test_blocks(),
     biomes: Biomes::test_blocks(),
   });
-  let generator = Arc::new(WorldBiomes::new(&context.blocks, &context.biomes, context.seed as u64));
+  let generator = Arc::new(WorldBiomes::new(&context.blocks, &context.biomes, context.seed));
   world.spawn_threads(&context, &generator);
 
   let mut chunk_pos = ChunkPos::new(0, 0);
@@ -36,11 +36,11 @@ fn bench_chunk(b: &mut Bencher) {
 #[bench]
 fn bench_base(b: &mut Bencher) {
   let context = Arc::new(rgen_world::Context {
-    seed:   1 as u64,
+    seed:   1_u64,
     blocks: Blocks::test_blocks(),
     biomes: Biomes::test_blocks(),
   });
-  let generator = Arc::new(WorldBiomes::new(&context.blocks, &context.biomes, context.seed as u64));
+  let generator = Arc::new(WorldBiomes::new(&context.blocks, &context.biomes, context.seed));
 
   let mut chunk_pos = ChunkPos::new(0, 0);
